worked on various parts of the project: implementing visit sorting by status, a month selection dialog in `VisitFragment`, and client search based on various parameters; enlarging the month-switching buttons in `VisitFragment`; and adding automatic address detection from coordinates when creating a visit.

+++ app/src/main/java/com/subefu/aquateka/model/data/entity/ClientEntity.kt
@@ -10,8 +10,8 @@ data class ClientEntity(
     val name: String,
     val phone: String,
     val address: String?,
-    val latitude: Float,
-    val longitude: Float,
+    val latitude: Double,
+    val longitude: Double,
     val periodMonth: Int?,
     val comment: String
 )
